+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Enable 'vars' key at endpoint node. [#328](https://github.com/scanapi/scanapi/pull/328)
 
 ## [2.4.0] - 2021-06-11
 ### Added